+++ NEWS.md
@@ -7,6 +7,9 @@ it biologically reasonable.
 that are present in only one sample when `distribution` is `neg.binom`. Instead,
 this function will use the empirical signature parameters from all cancer types.
 
+* Updated function `GenerateSyntheticTumors` to enable generating different number 
+of synthetic tumors for each cancer type.
+
 ### Added
 * Added default value 100 for argument `n.binom.size` in function `AddNoise`.